Moved scripts to Scripts file
