+++ UnityPlayer/UI/Assets/Scripts/PlayingTrackScript.cs
